requirements.txt
tensorflow==2.18.0
Pillow==10.4.0
opencv-python==4.10.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
